Add new API.
livebox_service_get_livebox_pkgname

To get the specified livebox pkgid of given appid (or livebox pkgid).
If the given id is the appid, the the function will returns livebox's
pkgid which is primary one.
and if the given id is the pkgid, it will just returns the pkgid.

Return pkgid is alloctated on the heap, or it will be NULL

(Validate the libexec)

Change-Id: I9aa7b3c83e705b8ee77a24dc3c02aadac536bcb9