Check the appinfo (pkgmgr), Replace the path
livebox_service_pkgname implementation is changed.

Try to get the livebox package name using given appid.
but if the given appid is not same with the "package" name.

Try to get the package name using pkgmgr API.
After it gets the "package" name, try to find the primary livebox again.

And return its result.
if it finds the primary livebox, the livebox_service_pkgname will
returns it after duplicate it on the heap.
or NULL returns.

Change-Id: I430581b35e6864a3d82a1987a17052468709e305