Add API for getting the "need_frame" attribute
Each box can get help from viewer to draw border for box contents.
In that case, the box can set "true" for need_frame attribute.

Then the viewer will draw border for it.
To get the need_frame value, this library supports livebox_service_need_frame API.
It will returns true of false about need_frame attribute.

If it said, true, the viewer can draw some border on its contents.
or the viewer should keep its original contents.

But it is just recomendataion, if the viewer draw something on the box,
There is no way to keep it in safe ;)

Change-Id: Ia54d2c592bbcd1252bfd9a35707c0ece2653f61e