Remove Product Header.
If the API is handles as PRODUCT Only Feature.
It must be removed from plastform scope.
So if it is not really intended to support only for the product module,
use the "@internal" mark for it instead of separate it to product module.
Then the API can be exists in platform side and it could be hidden from 3rd party developer.

Change-Id: Ie2060e43611cc2aa0c415849034bd8b18d79b4ef