Launch widget app with async api
When the widget-viewer launches widget app with sync api,
the widget app would be blocked in its elm_win_add function
because the widget-viewer is waiting for aul_launch_app api
and cannot process request of elm_win_add.

Note that this api is a temporary solution.
We need to investigate if there is a better solution.

Change-Id: I264b5bd3195ad9a204be406705d03eddaa3ed1cc
Signed-off-by: Semun Lee <sm79.lee@samsung.com>